requirements.txt
asgiref==3.8.1
Django==4.2.17
django-modeltranslation==0.19.12
pillow==11.1.0
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.12.2
